Modeling variability
One value is set for the whole model for the random fluctuation at a given time step in the resource base. The base level of the resource base is moved down on the Type 2 response curve so fluctuations have more of an effect and we're not stuck in the saturated part of the response. And, the call to negative_p_net is fixed to take the correct E argument, E_past.